changing to controllable char
